Added new key in epi and improved support to epi type file with list

+++ python3-epi/epimanager.py
@@ -273,7 +273,7 @@ class EpiManager:
 		try:
 			if self.epiFiles[order]["script"]["getStatus"]:
 				self.getStatus_byscript=True
-		except:
+		except Exception as e:
 			self.getStatus_byscript=False
 			pass
 
@@ -986,6 +986,9 @@ class EpiManager:
 		pkgs_ref=[]
 		result=False
 		content=""
+		pkgs={}
+		file_with_list=False	
+
 
 
 		if action=="install":
@@ -993,14 +996,24 @@ class EpiManager:
 			if epi_type == "file":
 				if self.token_result_install!="":
 					token=self.token_result_install[1]
+					if self.epi_conf["selection_enabled"]["active"]:
+						pkgs=self.epi_conf["pkg_list"]
+						file_with_list=True
 			
+
 			elif epi_type!="file":
 				pkgs=self.epi_conf["pkg_list"]
 		else:
 			epi_type=self.epiFiles[0]["type"]
 			if epi_type=="file":
 				token=self.token_result_remove[1]
-				
+				if self.epiFiles[0]["selection_enabled"]["active"]:
+					pkgs=self.epiFiles[0]["pkg_list"]
+					file_with_list=True
+					for item in pkgs:
+						if item["name"] in self.packages_selected:
+							pkgs_ref.append(item["name"])
+	
 			elif epi_type !="file":
 				pkgs=self.epiFiles[0]["pkg_list"]
 				for item in pkgs:
@@ -1008,19 +1021,20 @@ class EpiManager:
 						pkgs_ref.append(item["name"])
 
 
-		if epi_type=="file":
+		if epi_type=="file" and not file_with_list:
 			if os.path.exists(token):
 				file=open(token)
 				content=file.readline()
 				if '0' not in content:
 					result=False
+
 				else:
 					result=True	
 							
 				file.close()
 				os.remove(token)
 
-		elif epi_type !="file":
+		elif epi_type !="file" or file_with_list:
 			for item in pkgs:
 				if item["name"] in self.packages_selected:
 					if epi_type=="mix":
@@ -1030,8 +1044,11 @@ class EpiManager:
 							status=self.check_pkg_status(item["name"],0)
 			
 					else:
-						status=self.check_pkg_status(item["name"])
-	
+						if epi_type=="file":
+							status=self.check_pkg_status(item["name"],0)
+						else:
+							status=self.check_pkg_status(item["name"])
+
 				#if item["name"] in self.packages_selected:
 					dpkg_status[item["name"]]=status
 					if status!="installed":
@@ -1067,8 +1084,6 @@ class EpiManager:
 
 		return dpkg_status,result			
 
-		
-		
 	#def check_install_remove	
 
 	def postinstall_app(self):
@@ -1341,7 +1356,7 @@ class EpiManager:
 				zmd_status=1
 		else:
 			zmd_status=1
-			
+
 		return zmd_status
 	
 	#def get_zmd_status
